Projectile traces now implemented with dynamic beam simple niagara emitter setup in projectile blueprint. Also setting the owner of the projectiles to now be the owner of the pooler for the projectile. Changing the colour of the projectile trace depending on the owner either player or not. Fixed grappling with no weapon animation blending. Fixed janky tilting when locomotion with no weapon equipped.
New niagara emitter for projectile trace / bullet tracers which is just a dynamic beam emitter.

Emitter in the projectile actor and then just activating and deactivating when doing that in the projectile actor itself so not constantly having the emitter active when the projectile is not active.

The default pose for animation blueprint now layered blend per bone the idle animation at neck_01 so that the camera isn't janky and the tilting actually works properly when you don't have a weapon.